Use spring-cloud-stream to listen minio notifications
