+++ gridcapa-task-manager-app/src/test/java/com/farao_community/farao/gridcapa/task_manager/app/TaskManagerTest.java
@@ -29,7 +29,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 class TaskManagerTest {
 
     @MockBean
-    private TaskNotifier taskNotifier; // Useful to avoid AMQP connection that would fail
+    private TaskUpdateNotifier taskUpdateNotifier; // Useful to avoid AMQP connection that would fail
 
     @MockBean
     private MinioAdapter minioAdapter;
