+++ gridcapa-task-manager-app/src/test/java/com/farao_community/farao/gridcapa/task_manager/app/TaskProcessFileDeletionTest.java
@@ -31,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.*;
 class TaskProcessFileDeletionTest {
 
     @MockBean
-    private TaskNotifier taskNotifier; // Useful to avoid AMQP connection that would fail
+    private TaskUpdateNotifier taskUpdateNotifier; // Useful to avoid AMQP connection that would fail
 
     @Autowired
     private TaskRepository taskRepository;
